+++ R/int_addStrataCode.R
@@ -10,7 +10,7 @@
 
   # 'Declare' var
 
-  stratum = NULL
+  .stratum = NULL
 
   #
 
@@ -20,12 +20,25 @@
   data.table::setkeyv(strataOnly, names(strataOnly))
 
   uniqueStrata <- unique(strataOnly)
-  uniqueStrata[, stratum := (1 : .N)]
+  uniqueStrata[, .stratum := (1 : .N)]
 
   data.table::setkeyv(dtWork, names(strataOnly))
   dtWork <- uniqueStrata[dtWork]
 
   data.table::setkeyv(dtWork, key(dt))
 
-  dtWork
+  dtWork[]
+}
+
+# Stratified sample
+#
+# @param nrow Number of rows in the stratum
+# @param ncat Number of treatment categories
+# @return A sample draw from a stratum
+
+.stratSamp <- function(nrow, ncat) {
+  neach <- floor(nrow / ncat)
+  distrx <- rep(c(1:ncat), each = neach)
+  extra <- nrow - length(distrx)
+  sample(c(distrx, sample(1:ncat, extra)))
 }